discode/server/requirements.txt
discord.py
flask
flask_cors 